requirements.txt
python-telegram-bot==11.1.0
psycopg2-binary
environs
